config some code

+++ app/Views/index.php
@@ -7,11 +7,11 @@
 			<div class="col-8">
 				<div class="text-right">
 
-				<?php //if(session()->get('role') == 1):?>
+				<?php if(session()->get('role') == "manager"):?>
 					<a href="" class="btn btn-warning btn-sm text-white font-weight-bolder" data-toggle="modal" data-target="#createPizza">
 						<i class="material-icons float-left" data-toggle="tooltip" title="Add Pizza!" data-placement="left">add</i>&nbsp;Add
 					</a>
-				<?php //endif;?>
+				<?php endif;?>
 
 				</div>
 				<hr>
@@ -20,7 +20,7 @@
 						<th>Name</th>
 						<th>Price</th>
 						<th>Ingredients</th>
-						<?php if(session()->get('role') == 1):?>
+						<?php if(session()->get('role') == "manager"):?>
 							<th>Status</th>
 						<?php endif;?>
 					</tr>
@@ -30,10 +30,12 @@
 							<td ><?= $pizza['name']?></td>
 							<td ><?= $pizza['price']?></td>
 							<td><?= $pizza['ingredient']?></td>	
-							<td>	
-								<a href="/edit/<?= $pizza['id']?>" data-toggle="modal" data-target="#updatePizza"><i class="material-icons text-info" data-toggle="tooltip" title="Edit Pizza!" data-placement="left">edit</i></a>
-								<a href="/delete/<?= $pizza['id']?>" data-toggle="tooltip" title="Delete Pizza!" data-placement="right"><i class="material-icons text-danger">delete</i></a>
-							</td>
+							<?php if(session()->get('role') == "manager"):?>
+								<td>	
+									<a href="/edit/<?= $pizza['id']?>" data-toggle="modal" data-target="#updatePizza"><i class="material-icons text-info" data-toggle="tooltip" title="Edit Pizza!" data-placement="left">edit</i></a>
+									<a href="/delete/<?= $pizza['id']?>" data-toggle="tooltip" title="Delete Pizza!" data-placement="right"><i class="material-icons text-danger">delete</i></a>
+								</td>
+							<?php endif;?>
 						</tr>
 					<?php endforeach;?>
 				</table>
